# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "8조 API",
    "description": "Description"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Login",
      "description": "로그인"
    },
    {
      "name": "User",
      "description": "회원가입"
    },
    {
      "name": "Post",
      "description": "게시판"
    },
    {
      "name": "reply",
      "description": "댓글"
    }
  ],
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/posts/": {
      "get": {
        "tags": ["Post"],
        "description": "",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/posts/post": {
      "post": {
        "tags": ["Post"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/posts/modify/{postId}": {
      "put": {
        "tags": ["Post"],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/posts/delete/{postId}": {
      "patch": {
        "tags": ["Post"],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/posts/{postId}": {
      "get": {
        "tags": ["Post"],
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/replies/": {
      "get": {
        "tags": ["reply"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "postId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/replies/post": {
      "post": {
        "tags": ["reply"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "replyContent": {
                  "example": "any"
                },
                "postId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/replies/delete": {
      "patch": {
        "tags": ["reply"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "postId": {
                  "example": "any"
                },
                "replyId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/replies/modify": {
      "put": {
        "tags": ["reply"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "postId": {
                  "example": "any"
                },
                "replyId": {
                  "example": "any"
                },
                "replyContent": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/register/": {
      "post": {
        "tags": ["User"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "userPw": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/login/": {
      "post": {
        "tags": ["Login"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "userPw": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/user/me": {
      "get": {
        "tags": ["Login"],
        "description": "",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
    "Users": {
      "type": "object",
      "properties": {
        "id": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "Number"
            }
          }
        },
        "userId": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "nickname": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "userPw": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "salt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "createdAt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "updatedAt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        }
      }
    },
    "Posts": {
      "type": "object",
      "properties": {
        "postId": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "Number"
            }
          }
        },
        "userId": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "title": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "content": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "image": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "postDelType": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "createdAt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "updatedAt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        }
      }
    },
    "Replies": {
      "type": "object",
      "properties": {
        "replyId": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "Number"
            }
          }
        },
        "userId": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "postId": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "replyContent": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "replyDelType": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "createdAt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "updatedAt": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "String"
            }
          }
        }
      }
    }
  }
}